requirements.txt
# Database
oracledb==2.0.1

# GUI
PyQt5==5.15.10

# Computer Vision
opencv-python==4.8.1.78
imutils==0.5.4

# Machine Learning / Deep Learning
tensorflow-cpu==2.16.1
scikit-learn==1.3.2

# Input Control
keyboard==0.13.5
pyautogui==0.9.54

# Web Framework
Flask==3.0.0
requests==2.31.0

# Image Processing
Pillow==10.1.0

# System Monitoring (Windows)
WMI==1.5.1
pywin32==306

# Scientific Computing
numpy==1.26.2
